+++ CHANGELOG.md
@@ -2,6 +2,36 @@
 
 This is a list of manually mantained changes and updates for each version.
 
+## Version 3.5.29
+
+## Added
+
+None
+
+## Changed
+
+#### RPC state object added a function respond(), which is the same as callback
+
+#### Dependency gracelog version updated
+
+to v0.6.7
+
+## Deprecated
+
+#### gracenode.registerShutdownTask()
+
+#### gracenode.router
+
+Please use `gracenode.http` instead.
+
+#### gracenode.lib.cloneObj()
+
+## Removed
+
+None
+
+***
+
 ## Version 3.5.28
 
 ## Added